move memberModal and paymentModal to components folder
